Final build fix

+++ source/backend/plugin/LadspaPlugin.cpp
@@ -1268,7 +1268,7 @@ public:
 
     void sampleRateChanged(const double newSampleRate) override
     {
-        CARLA_ASSERT_INT(newSampleRate > 0.0, newSampleRate);
+        CARLA_ASSERT_INT(newSampleRate > 0.0, (int)newSampleRate);
         carla_debug("LadspaPlugin::sampleRateChanged(%g) - start", newSampleRate);
 
         // TODO
